refactor: Relocate End Game button beside progress bar with responsive layout
I refactored the "End Game" button and its confirmation dialog into a reusable `EndGameButton.tsx` component.

I introduced a new `ProgressAndControlsSection.tsx` component to manage the layout of the `WordProgressBar` and the `EndGameButton`.
This new section arranges these two components:
- Inline on desktop screens (progress bar left, button right).
- Stacked on mobile screens (progress bar top, button bottom).

I updated `WordDisplaySection.tsx` to use this new `ProgressAndControlsSection`.
`GameHeader.tsx` no longer contains the "End Game" button.
I also performed minor cleanup in `app/page.tsx`.

The functionality of the "End Game" confirmation dialog and the progress bar remains unchanged.
